+++ packages/headless/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.7.1](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.7.0...@coveo/headless@2.7.1) (2023-01-25)
+
+**Note:** Version bump only for package @coveo/headless
+
 # [2.7.0](https://github.com/coveo/ui-kit/compare/@coveo/headless@2.6.1...@coveo/headless@2.7.0) (2023-01-23)
 
 ### Features